+++ include/timestepper.hpp
@@ -54,7 +54,7 @@ class TimeStepper{
     virtual ~TimeStepper(); 
 
     /* Return the state at a certain time index */
-    Vec getState(int tindex);
+    Vec getState(size_t tindex);
 
     /* Solve the ODE forward in time with initial condition rho_t0. Return state at final time step */
     Vec solveODE(int initid, Vec rho_t0);
